# Avoid temporary vectors and maps during KT stabilisation