Creating and destroying the vectors and maps was slowing
numerical stabilisation.

Here are some test timing before and after this PR. 

| test  | before | after | speedup |
| ------------- | ------------- | ------------- | ------------- |
| `flux_limited_TVD_advection/fltvd_3D.i` | 36.763 | 21.703 | 1.69 |
| `flux_limited_TVD_pflow/pffltvd_3D.i` | 41.762 | 17.89 | 2.33 |
| `flux_limited_TVD_pflow/pffltvd_2D_trimesh.i` | 520.612 | 377.17 | 1.38 |

Refs #13857